requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pymongo==4.13.2
groq==0.30.0
numpy==2.2.0
pypdf==5.8.0
Pillow==11.3.0
pytesseract==0.3.13
opencv-python-headless==4.12.0.88
easyocr==1.7.2
python-dotenv==1.1.1
redis==5.0.8
